Refactoring various scripts
Generally getting rid of initial values made in the class declarations and putting them into the start function which will often be in a initialsetup type function. When an option to edit the variable in the editor is present (with SerializeField) the value is being set there rather than in the script.

Also removed any other unnecessary print statements, uses of SerializeField etc.